Feat: get all permissions

+++ app/core/permissions/endpoints_permissions.py
@@ -6,7 +6,7 @@ Group management is part of the core of Hyperion. These endpoints allow managing
 
 import logging
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.groups.groups_type import GroupType
@@ -15,7 +15,9 @@ from app.dependencies import (
     get_db,
     is_user_in,
 )
+from app.module import full_name_permissions_list, permissions_list
 from app.types.module import CoreModule
+from app.utils.tools import is_group_id_valid
 
 router = APIRouter(tags=["Groups"])
 
@@ -26,6 +28,21 @@ core_module = CoreModule(
 )
 
 hyperion_security_logger = logging.getLogger("hyperion.security")
+
+
+@router.get(
+    "/permissions/list",
+    response_model=list[str],
+    status_code=200,
+)
+async def read_permissions_list(
+    user=Depends(is_user_in(GroupType.admin)),
+):
+    """
+    Return all permissions from database
+    """
+
+    return full_name_permissions_list
 
 
 @router.get(
@@ -76,6 +93,16 @@ async def create_permission(
     """
     Create a new permission in database
     """
+    if not await is_group_id_valid(permission.group_id, db):
+        raise HTTPException(
+            status_code=404,
+            detail="Group not found",
+        )
+    if permission.permission_name not in permissions_list:
+        raise HTTPException(
+            status_code=404,
+            detail="Permission not found",
+        )
 
     await cruds_permissions.create_permission(permission, db)
     return {"message": "Permission created successfully"}
@@ -93,6 +120,18 @@ async def delete_permission(
     """
     Delete a permission from database by name
     """
+    permission_db = (
+        await cruds_permissions.get_permissions_by_group_id_and_permission_name(
+            db,
+            permission.permission_name,
+            permission.group_id,
+        )
+    )
+    if not permission_db:
+        raise HTTPException(
+            status_code=404,
+            detail="Permission not found",
+        )
 
     await cruds_permissions.delete_permission(db, permission)
     return {"message": "Permission deleted successfully"}
